requirements.txt
Flask==2.3.3
flask-sock==0.7.0
python-chess==1.10.0
Werkzeug==2.3.7
